@@ -22,49 +22,44 @@
     "email": "support@digitalbazaar.com"
   },
   "license": "BSD-3-Clause",
-  "main": "lib/index.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "files": [
     "lib/**/*.js"
   ],
   "dependencies": {
     "credentials-context": "^2.0.0",
     "jsonld": "^5.2.0",
-    "jsonld-signatures": "^9.3.0"
+    "jsonld-signatures": "^9.3.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.8",
-    "@babel/plugin-transform-modules-commonjs": "^7.13.8",
-    "@babel/plugin-transform-runtime": "^7.13.9",
-    "@babel/preset-env": "^7.13.9",
-    "@babel/runtime": "^7.13.9",
-    "@digitalbazaar/ed25519-signature-2018": "^2.0.1",
-    "@digitalbazaar/ed25519-signature-2020": "^3.0.0",
-    "@digitalbazaar/ed25519-verification-key-2018": "^3.1.1",
-    "@digitalbazaar/ed25519-verification-key-2020": "^3.1.0",
-    "babel-loader": "^8.2.2",
-    "chai": "^4.3.3",
+    "@digitalbazaar/ed25519-signature-2018": "^3.0.0",
+    "@digitalbazaar/ed25519-signature-2020": "^4.0.1",
+    "@digitalbazaar/ed25519-verification-key-2018": "^4.0.0",
+    "@digitalbazaar/ed25519-verification-key-2020": "^4.0.0",
+    "c8": "^7.11.3",
+    "chai": "^4.3.6",
     "cross-env": "^7.0.3",
     "did-context": "^3.1.1",
-    "did-veres-one": "^13.0.0",
-    "eslint": "^7.21.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.2.0",
-    "karma": "^6.1.1",
-    "karma-babel-preprocessor": "^8.0.1",
+    "did-veres-one": "^15.0.0",
+    "eslint": "^8.15.0",
+    "eslint-config-digitalbazaar": "^3.0.0",
+    "eslint-plugin-jsdoc": "^39.2.9",
+    "eslint-plugin-unicorn": "^42.0.0",
+    "karma": "^6.3.20",
     "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-sourcemap-loader": "^0.3.8",
     "karma-webpack": "^5.0.0",
-    "mocha": "^8.3.1",
+    "mocha": "^10.0.0",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^15.1.0",
     "uuid": "^8.3.2",
     "veres-one-context": "^12.0.0",
-    "webpack": "^5.24.3"
+    "webpack": "^5.72.1"
   },
-  "nyc": {
+  "c8": {
     "exclude": [
       "test"
     ],
@@ -73,13 +68,8 @@
       "text-summary"
     ]
   },
-  "browser": {
-    "Buffer": false,
-    "crypto": false,
-    "util": false
-  },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "keywords": [
     "JSON",
@@ -93,10 +83,10 @@
   "scripts": {
     "test": "npm run lint && npm run test-node && npm run test-karma",
     "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 10000 test/*.spec.js",
-    "test-karma": "karma start karma.conf.js",
-    "lint": "eslint lib test/*.spec.js",
-    "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
-    "coverage-ci": "cross-env NODE_ENV=test nyc --reporter=lcovonly npm run test-node",
-    "coverage-report": "nyc report"
+    "test-karma": "karma start karma.conf.cjs",
+    "lint": "eslint .",
+    "coverage": "cross-env NODE_ENV=test c8 --reporter=lcov --reporter=text-summary --reporter=text npm run test-node",
+    "coverage-ci": "cross-env NODE_ENV=test c8 --reporter=lcovonly --reporter=text-summary --reporter=text npm run test-node",
+    "coverage-report": "c8 report"
   }
 }
