@@ -1,10 +1,10 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
 
+export const CONTEXT_URL = 'https://www.w3.org/ns/odrl.jsonld';
 /* eslint-disable quote-props, key-spacing, max-len */
-module.exports = {
+export const CONTEXT = {
   '@context': {
     'odrl':    'http://www.w3.org/ns/odrl/2/',
     'rdf':     'http://www.w3.org/1999/02/22-rdf-syntax-ns#',
