@@ -1,8 +1,7 @@
-const assertionController = {
+export const assertionController = {
   '@context': 'https://w3id.org/security/v2',
   id: 'https://example.edu/issuers/565049',
   // actual keys are going to be added in the test suite before() block
   assertionMethod: [],
   authentication: []
 };
-module.exports = assertionController;
